requirements.txt
discord.py
aiosqlite
shortuuid
Flask[async]
Flask-Cors
aiohttp
cogwatch
openai
